+++ CHANGELOG.md
@@ -9,7 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-*
+* Support for the list and get purchase operations in `PurchaseAPI`
+* Support for the list and get consignment operations in `ConsignmentAPI`
 
 ### Changed